pysmu: use get_device to simplify acquiring a matching device
ValueError is raised if no matching device is found.

Also, note that an int is used for the device ID since python doesn't do
overflow checking when parsing args for 'I' and other unsigned types. So
we just use an int and error out if a negative value is passed later.

The other way this is often handled is writing custom convertor methods
that do proper overflow checking during type conversion.