Device: write(): add condition variable waiting for new buffer submissions
This allows the write threads to fully idle when unused.

Also, we now throw errors if a write() call takes longer than a certain
amount of time to be submitted to the thread to help alleviate hanging
situations. Currently this is set to a static 100ms, but could be
switched to be dependent on the submitted buffer size.