Device::encode_out(): add support to use the next value in the queue without discarding it
This helps smooth the transitions across multiple non-continuous reads
of continuous waveforms. In effect, it primes the hardware to know what
value is coming so when the actual sample value is written and then read
it should be closer to what is expected for the transition sample
values (start/end of non-continuous sessions).