bindings/python/setup.py: Always regenerate cython extensions.
The cython file is generated in the first build. Any modification done in
the source files (libsmu.pyx for example) is not propagated to the
python bindings. In order to do so, the libsmu.cpp (generated) file
should be removed after every build.
To fix this, we configure setup.py to always regenerate the extension files.

Signed-off-by: Alexandra.Trifan <Alexandra.Trifan@analog.com>