Session: Modify the scanning process to avoid having wrong captured samples.
The scan() method is used to detect if any devices were connected or
disconnected. In order to identify them, the serial number is extracted
using a control_transfer function. If the control_transfer is called
while a device is already capturing, some wrong values will appear in the
capture. To avoid this, we use the serial number when the device if added
for the first time, then we identify devices using a combination of
USB bus_id and USB address. This way we can detect if a running device
was disconnected (without interfering with the signal)

Signed-off-by: Alexandra Trifan <Alexandra.Trifan@analog.com>