bin/pysmu: Fix urllib imports for python2 and python3 compatibility
Submodule urlretrieve is not provided anymore by urllib but
urllib.request in python3 ; and module urllib2 is not available in
python3.

The import is fixed by calling the appropriate module for python3.
The import is made compatible for both python2 and python3.

The actual call urlopen() in pysmu.flash() method is also fixed
according to the new import implementation.

Signed-off-by: Cyber-G <cygerpham@free.fr>